adding frontend audit
